Comments and additional tests
